+++ Gemfile
@@ -1,4 +1,4 @@
 source 'https://rubygems.org'
 
-# Specify your gem's dependencies in mergometer.gemspec
+# Specify your gem's dependencies in hub_link.gemspec
 gemspec
