--- exe/mergometer
@@ -1,6 +1,0 @@
-#!/usr/bin/env ruby
-
-require "mergometer"
-require "mergometer/cli"
-
-Mergometer::Cli.run(ARGV)
