+++ lib/hub_link.rb
@@ -1,0 +1,8 @@
+require "dotenv/load"
+require "hub_link/configuration"
+require "hub_link/stream"
+require "hub_link/version"
+
+module HubLink
+  Configuration.new(cache_time: 72 * 3600).apply
+end
