+++ lib/hub_link/configuration.rb
@@ -2,7 +2,7 @@ require "faraday/detailed_logger"
 require "faraday_middleware"
 require "active_support"
 
-module Mergometer
+module HubLink
   class Configuration
     def initialize(cache_time: 3600)
       @cache_time = cache_time
