+++ lib/hub_link/core_ext/float.rb
