+++ lib/hub_link/version.rb
@@ -1,3 +1,3 @@
-module Mergometer
+module HubLink
   VERSION = "0.1.0"
 end
