--- lib/mergometer/cli.rb
@@ -1,32 +1,0 @@
-module Mergometer
-  class Cli
-    def self.run(*args)
-      new(*args).run
-    end
-
-    def initialize(argv)
-      @argv = argv
-    end
-
-    def run
-      generate_reports
-    end
-
-    private
-
-      attr_reader :argv
-
-      def generate_reports
-        ReportsGenerator.new(repo_names).run
-      end
-
-      def repo_names
-        argv[0].presence || stop
-      end
-
-      def stop
-        puts "\nUsage: OCTOKIT_ACCESS_TOKEN=<token> #{$0} <github_org/repo_name>"
-        exit
-      end
-  end
-end
