+++ test/hub_link_test.rb
@@ -1,8 +1,8 @@
 require 'test_helper'
 
-class MergometerTest < Minitest::Test
+class HubLinkTest < Minitest::Test
   def test_that_it_has_a_version_number
-    refute_nil ::Mergometer::VERSION
+    refute_nil ::HubLink::VERSION
   end
 
   def test_it_does_something_useful
