+++ test/test_helper.rb
@@ -1,4 +1,4 @@
 $LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
-require 'mergometer'
+require 'hub_link'
 
 require 'minitest/autorun'
